pyproject.toml
[project]
name = "django-first-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
